HTTP app version and name header
