+++ component/http/v2/router/httprouter/router_test.go
@@ -2,9 +2,9 @@ package httprouter
 
 import (
 	"net/http"
+	"net/http/httptest"
 	"testing"
 
-	"github.com/beatlabs/patron/component/http/middleware"
 	v2 "github.com/beatlabs/patron/component/http/v2"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -40,156 +40,64 @@ func TestNew(t *testing.T) {
 	}
 }
 
-func TestRoutes(t *testing.T) {
+// TODO: proper router tests. using httptest to start the httprouter and try out all routes plus a custom one.
+
+func TestVerifyRouter(t *testing.T) {
 	t.Parallel()
-	type args struct {
-		routes []*v2.Route
-	}
-	tests := map[string]struct {
-		args        args
-		expectedErr string
-	}{
-		"success": {args: args{routes: []*v2.Route{{}, {}}}},
-		"fail":    {args: args{routes: nil}, expectedErr: "routes are empty"},
-	}
-	for name, tt := range tests {
-		tt := tt
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			cfg := &Config{}
-			err := Routes(tt.args.routes...)(cfg)
-			if tt.expectedErr != "" {
-				assert.EqualError(t, err, tt.expectedErr)
-			} else {
-				assert.Equal(t, tt.args.routes, cfg.routes)
-			}
-		})
-	}
-}
 
-func TestAliveCheck(t *testing.T) {
-	t.Parallel()
-	type args struct {
-		acf v2.LivenessCheckFunc
-	}
-	tests := map[string]struct {
-		args        args
-		expectedErr string
-	}{
-		"success": {args: args{acf: func() v2.AliveStatus { return v2.Alive }}},
-		"fail":    {args: args{acf: nil}, expectedErr: "alive check function is nil"},
-	}
-	for name, tt := range tests {
-		tt := tt
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			cfg := &Config{}
-			err := AliveCheck(tt.args.acf)(cfg)
-			if tt.expectedErr != "" {
-				assert.EqualError(t, err, tt.expectedErr)
-			} else {
-				assert.NotNil(t, cfg.aliveCheckFunc)
-			}
-		})
-	}
-}
+	route, err := v2.NewRoute(http.MethodGet, "/api/", func(writer http.ResponseWriter, request *http.Request) {
+		writer.WriteHeader(200)
+	})
+	require.NoError(t, err)
 
-func TestReadyCheck(t *testing.T) {
-	t.Parallel()
-	type args struct {
-		rcf v2.ReadyCheckFunc
-	}
-	tests := map[string]struct {
-		args        args
-		expectedErr string
-	}{
-		"success": {args: args{rcf: func() v2.ReadyStatus { return v2.Ready }}},
-		"fail":    {args: args{rcf: nil}, expectedErr: "ready check function is nil"},
-	}
-	for name, tt := range tests {
-		tt := tt
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			cfg := &Config{}
-			err := ReadyCheck(tt.args.rcf)(cfg)
-			if tt.expectedErr != "" {
-				assert.EqualError(t, err, tt.expectedErr)
-			} else {
-				assert.NotNil(t, cfg.readyCheckFunc)
-			}
-		})
-	}
-}
+	appName := "appName"
+	appVersion := "1.1"
+	appVersionHeader := "X-App-Version"
+	appNameHeader := "X-App-Name"
 
-func TestDeflateLevel(t *testing.T) {
-	t.Parallel()
-	cfg := &Config{}
-	err := DeflateLevel(10)(cfg)
-	assert.NoError(t, err)
-	assert.Equal(t, 10, cfg.deflateLevel)
-}
+	router, err := New(Routes(route), EnableAppNameHeaders(appName, appVersion))
+	require.NoError(t, err)
 
-func TestMiddlewares(t *testing.T) {
-	t.Parallel()
-	type args struct {
-		mm []middleware.Func
-	}
-	tests := map[string]struct {
-		args        args
-		expectedErr string
-	}{
-		"success": {args: args{mm: []middleware.Func{func(next http.Handler) http.Handler { return next }}}},
-		"fail":    {args: args{mm: nil}, expectedErr: "middlewares are empty"},
-	}
-	for name, tt := range tests {
-		tt := tt
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-			cfg := &Config{}
-			err := Middlewares(tt.args.mm...)(cfg)
-			if tt.expectedErr != "" {
-				assert.EqualError(t, err, tt.expectedErr)
-			} else {
-				assert.Len(t, cfg.middlewares, 1)
-			}
-		})
-	}
-}
+	srv := httptest.NewServer(router)
+	defer func() {
+		srv.Close()
+	}()
 
-func TestDisableProfiling(t *testing.T) {
-	t.Parallel()
-	cfg := &Config{}
-	err := EnableExpVarProfiling()(cfg)
-	assert.NoError(t, err)
-	assert.True(t, cfg.enableProfilingExpVar)
-}
+	assertResponse := func(t *testing.T, rsp *http.Response) {
+		assert.Equal(t, http.StatusOK, rsp.StatusCode)
+		assert.Equal(t, appName, rsp.Header.Get(appNameHeader))
+		assert.Equal(t, appVersion, rsp.Header.Get(appVersionHeader))
+	}
 
-func TestEnableAppNameHeaders(t *testing.T) {
-	type args struct {
-		name    string
-		version string
-	}
-	tests := map[string]struct {
-		args        args
-		expectedErr string
-	}{
-		"success":         {args: args{name: "name", version: "version"}},
-		"missing name":    {args: args{name: "", version: "version"}, expectedErr: "app name was not provided"},
-		"missing version": {args: args{name: "name", version: ""}, expectedErr: "app version was not provided"},
-	}
-	for name, tt := range tests {
-		tt := tt
-		t.Run(name, func(t *testing.T) {
-			cfg := &Config{}
-			err := EnableAppNameHeaders(tt.args.name, tt.args.version)(cfg)
+	t.Run("check metrics endpoint", func(t *testing.T) {
+		rsp, err := http.Get(srv.URL + "/metrics")
+		require.NoError(t, err)
+		assertResponse(t, rsp)
+	})
 
-			if tt.expectedErr != "" {
-				assert.EqualError(t, err, tt.expectedErr)
-				assert.Nil(t, cfg.appNameVersionMiddleware)
-			} else {
-				assert.NoError(t, err)
-				assert.NotNil(t, cfg.appNameVersionMiddleware)
-			}
-		})
-	}
+	t.Run("check alive endpoint", func(t *testing.T) {
+		rsp, err := http.Get(srv.URL + "/alive")
+		require.NoError(t, err)
+		assertResponse(t, rsp)
+	})
+
+	t.Run("check alive endpoint", func(t *testing.T) {
+		rsp, err := http.Get(srv.URL + "/ready")
+		require.NoError(t, err)
+		assertResponse(t, rsp)
+	})
+
+	t.Run("check pprof endpoint", func(t *testing.T) {
+		rsp, err := http.Get(srv.URL + "/debug/pprof")
+		require.NoError(t, err)
+		assertResponse(t, rsp)
+	})
+
+	t.Run("check provided endpoint", func(t *testing.T) {
+		rsp, err := http.Get(srv.URL + "/api")
+		require.NoError(t, err)
+		assertResponse(t, rsp)
+	})
+
+	
 }
